NFT Data Science Project
I would like to look at the largest marketplace OpenSea. They have the following categories of NFTs:
- Art
- Collectibles
- Domain Names
- Music
- Photography
- Sports
- Trading Cards
- Utility
- Virtual Worlds


We can perform an analysis of each category and gather characteristics that focus on two things
- The NFT itself
- The NFT creator

The goal would be to utilize a dependent variable of successful sale for an item within the collection or the collection as a whole. We could also measure the speed to sale (time from minting to the first sale) and the volume of sales. 

For independent variables we could use the following:
- characteristics of the NFT and the NFT creators (described in the following pages)
- Twitter activity (amount of tweets, retweets)
- Overall Frequency
- Prelaunch activity (prior to the NFT being available)
- Do they announce that it is coming soon?
- Postlaunch activity (once the NFT becomes available)
- Sentiment 
- Visibility (measured by the amount of external links associated to the NFT and the NFT creator)
- Type of cryptocurrency

Once we have gathered the data we could then identify the traits that are associated with successful sales of NFTs. 


When we look at an NFT on OpenSea you can see various characteristics, they include the following:
- NFT name
-NFT creator
- NFT description
- Links to external sites 
- External website
- Discord
- Instagram
- Twitter
- Items in the collection
- Activity
- Date of creation
- Date of first sale

You can also identify characteristics of the NFT creator, they include the following:
- Creator name
- Description
- Date joined 
- External Links
- Collected
- Have they purchased other NFTs
- Created
- Favorited
- Activity